feat: reply 요청 기능 개선 및 코드 분리
