+++ apps/web/src/shared/ui/index.ts
@@ -1,2 +1,1 @@
-export { OutLineButton } from '@/src/shared/ui/outline-button';
 export { useInfiniteScroll } from '@/src/shared/ui/useInfiniteScroll';
