--- apps/web/src/shared/ui/outline-button.tsx
@@ -1,14 +1,0 @@
-import { Button } from '@workspace/ui/components/button';
-export function OutLineButton({
-  text,
-  action,
-}: {
-  text: string;
-  action: () => void;
-}) {
-  return (
-    <Button variant={'outline'} onClick={action}>
-      {text}
-    </Button>
-  );
-}
